% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_content.R
\name{render_exercises}
\alias{render_exercises}
\title{Renderizado de módulos en formato HTML}
\usage{
render_exercises(ejercicio, resuelto = FALSE)
}
\arguments{
\item{ejercicio}{caracter indicando cuál ejercicio visualizar:
'Ejercicio_practico1.1', 'Ejercicio_practico1.2', 'Ejercicio_practico2', or 'Ejercicio_practico3'}

\item{resuelto}{argumento lógico. Se debe visualizar la solución? FALSE por defecto}
}
\value{
esta función lleva al usiario al ejercicio seleccionado.
}
\description{
Visualizar ejercicios en formato HTML
}
