% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_content.R
\name{render_lectures}
\alias{render_lectures}
\title{Visualizar módulos en formato HTML}
\usage{
render_lectures(modulo)
}
\arguments{
\item{modulo}{caracter indicando cuál módulo visualizar:
'Intro'  : Introducción al curso
'Modulo1': Introducción a R - Parte 1
'Modulo2': Introducción a R - Parte 2
'Modulo3': Rásters y datos espaciales
'Modulo4': Procesamiento de datos
'Modulo5' : Estadísticas espaciales y temporales
'Modulo6': Evaluación de productos de precipitación
'Modulo7': Introducción a machine learning - I
'Modulo8': Introducción a machine learning - II
'Modulo9': Fusión de productos de precipitación
'Modulo10': Funciones
'Modulo11': Cómputo en paralelo
'Modulo12': Trabajando con RMarkdown y Leaflet
'Modulo13': Descarga Automática de Datos"}
}
\value{
esta función lleva al usiario al módulo seleccionado.
}
\description{
Visualizar módulos en formato HTML
}
